requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0

# CCXT Pro (now open source!)
ccxt>=4.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# HTTP client
aiohttp>=3.9.0

# Authentication
bcrypt>=4.1.0
python-multipart>=0.0.6

# Templating
jinja2>=3.1.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.26.0

# Discord bot (for message cleanup)
discord.py>=2.3.0

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
